Fix bug with pretty printing OutputVar
If a `Outputvar` is empty, then there was an issue with computing the
amount of padding to add. This can be fixed by adding `init = 0` when
taking the maximum, so the amount of padding is 0 if there is nothing to
pad. The edge case of the value of the key-value pairs in
`dim_attributes` being something that is not a dictionary is handled by
printing it instead of trying to iterate over it. Lastly, if there are
0 elements in a dimension, it print "0 element" instead of just "0".